Added Ajax functionality for delete

+++ delete.php
@@ -1,6 +1,15 @@
 <?php
+// No note variable found so exit
+if ( ! isset( $_GET['delete_id'] ) ) {
+	die();
+}
+
 require 'connection.php';
-$sql = $conn->prepare( 'DELETE  FROM notes WHERE id=?' );
-$sql->bind_param( 'i', $_GET['deleteid'] );
-$sql->execute();
-header( 'location:show.php' );
+
+$sql = $conn->prepare( 'DELETE FROM notes WHERE id=?' );
+$sql->bind_param( 'i', $_GET['delete_id'] );
+
+if ( $sql->execute() ) {
+	echo 1;
+	die();
+}
